+++ src/main/java/by/brel/service/impl/CustomerServiceImpl.java
@@ -51,7 +51,7 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     @Transactional
-    public long getCountAllCustomers() {
+    public int getCountAllCustomers() {
         return customerDAO.getCountAllCustomers();
     }
 }
